requirements.txt
Flask==0.10.1
SQLAlchemy>=1.0.3
sqlalchemy-gevent>=0.2
redis==2.10.5
celery==4.0.0
Flask-Login==0.3.2
Flask-Migrate==1.4.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-Session==0.2
Flask-Cache==0.13.1
Flask-WTF==0.11
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
WTForms==2.0.2
Werkzeug==0.9.6
alembic==0.7.6
amqp==2.1.1
anyjson==0.3.3
ujson==1.34
simplejson==3.8.1
argparse==1.2.1
#billiard==3.3.0.20
billiard==3.5.0.2
blinker==1.3
itsdangerous==0.24
kombu>=3.0.30
mysql-connector-python-rf==2.1.3
python-gcm==0.1.5
apns==2.0.1
pytz==2015.4
wsgiref==0.1.2
xlrd==0.9.3
openpyxl==2.3.0-b1
#MySQL-python==1.2.5
#umysqldb==1.0.3
#pyMySQL>=0.7.5
#git+git://github.com/erynet/umysqldb.git@develop
gevent==1.1.2
pyasn1==0.1.9
pyopenssl==0.15.1
ndg-httpsclient==0.4.0
requests==2.9.1
#python-redis-lock==2.3.0
PyCrypto==2.6.1
pylru==1.0.9
python-daemon==2.1.1
raven[flask]
virtualenv